[refactor]: style 및 컴포넌트 폴더구조 변경 및 세분화
