+++ src/app/new-post/page.tsx
@@ -1,12 +1,9 @@
-import Footer from '@/components/common/Footer';
-import Header from '@/components/common/Header';
 import NewPost from '@/components/new-post/NewPost';
 import style from '@/styles/new-post/new-post-page.module.css';
 
 function NewPostPage() {
   return (
     <div className={style.new_post_page}>
-      <Header></Header>
       <NewPost></NewPost>
     </div>
   );
