+++ src/styles/new-post/banner/banner.module.css
@@ -1,0 +1,11 @@
+.banner {
+  width: calc(100% - 60px);
+  height: 25%;
+
+  padding: 15px 30px;
+
+  display: flex;
+  justify-content: space-between;
+
+  background: #ffe63d;
+}
