+++ src/styles/new-post/banner/post-upload.module.css
@@ -1,59 +1,3 @@
-.banner {
-  width: calc(100% - 60px);
-  height: 25%;
-
-  padding: 15px 30px;
-
-  display: flex;
-  justify-content: space-between;
-
-  background: #ffe63d;
-}
-
-.info {
-  display: flex;
-  flex: 1;
-  align-items: flex-start;
-  flex-direction: column;
-  gap: 5px;
-}
-
-.title {
-  color: #000;
-  font-size: 50px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: normal;
-
-  border: 0;
-  background: none;
-  outline: none;
-}
-
-.title::placeholder {
-  color: #000;
-}
-
-.desc {
-  width: 70%;
-  height: 100%;
-
-  color: #000;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-
-  border: 0;
-  background: none;
-  outline: none;
-  resize: none;
-}
-
-.desc::placeholder {
-  color: #000;
-}
-
 .interaction_section {
   display: flex;
   flex: 1;
