--- src/styles/new-post/main/new-post.module.css
@@ -1,5 +1,0 @@
-.new_post {
-  width: 1250px;
-  margin: 0 auto;
-  padding-bottom: 50px;
-}
