+++ src/styles/new-post/new-post-page.module.css
@@ -1,4 +1,9 @@
 .new_post_page {
   width: 100%;
-  height: 100%;
+}
+
+.new_post {
+  width: 1250px;
+  margin: 0 auto;
+  padding-bottom: 50px;
 }
