--- src/styles/posts/banner.module.css
@@ -1,123 +1,0 @@
-.banner {
-  width: 100%;
-
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.list_container {
-  padding: 60px 60px;
-
-  display: flex;
-  flex-direction: column;
-  align-items: flex-end;
-  gap: 13px;
-
-  background: #ffe63d;
-}
-
-.list_header {
-  width: 595px;
-
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-
-  border-bottom: 1px solid rgba(0, 0, 0, 0.15);
-}
-
-.category_wrapper {
-  display: flex;
-  align-items: center;
-  gap: 5px;
-}
-
-.category_name {
-  color: #000;
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: normal;
-}
-
-.post_num {
-  color: #000;
-  font-size: 10px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-}
-
-.pagination_wrapper {
-  display: flex;
-  gap: 14px;
-}
-
-.pagination_item {
-  color: #8a8a8a;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-
-  cursor: pointer;
-}
-
-.post_list {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.post_list_item {
-  width: 650px;
-
-  display: flex;
-  justify-content: space-between;
-}
-
-.order_title {
-  display: flex;
-  gap: 33px;
-}
-
-.order {
-  color: #000;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-}
-
-.title {
-  color: #000;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-
-  cursor: pointer;
-}
-
-.date {
-  color: #8a8a8a;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-}
-
-.active_title {
-  text-decoration: underline;
-  font-weight: 700;
-}
-
-.active_order {
-  font-weight: 700;
-}
-
-.active_pagination_item {
-  color: #000;
-  text-decoration: underline;
-}
