+++ src/styles/posts/content/content.module.css
@@ -1,0 +1,5 @@
+.content_container {
+  width: 100%;
+
+  display: flex;
+}
