+++ src/styles/posts/content/detail.module.css
@@ -1,0 +1,30 @@
+.post_content {
+  width: 770px;
+
+  padding-top: 80px;
+  margin: 0 auto 100px auto;
+}
+
+.post_info {
+  display: flex;
+  flex-direction: column;
+  gap: 15px;
+  margin-bottom: 80px;
+}
+
+.post_title {
+  color: #1d1ac7;
+  font-size: 45px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: normal;
+}
+
+.post_date,
+.post_author {
+  color: #000;
+  font-size: 14px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: normal;
+}
