+++ src/styles/posts/post-list/post-list-header.module.css
@@ -1,0 +1,51 @@
+.list_header {
+  width: 595px;
+
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+
+  border-bottom: 1px solid rgba(0, 0, 0, 0.15);
+}
+
+.category_wrapper {
+  display: flex;
+  align-items: center;
+  gap: 5px;
+}
+
+.category_name {
+  color: #000;
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: normal;
+}
+
+.post_num {
+  color: #000;
+  font-size: 10px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: normal;
+}
+
+.pagination_wrapper {
+  display: flex;
+  gap: 14px;
+}
+
+.pagination_item {
+  color: #8a8a8a;
+  font-size: 12px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: normal;
+
+  cursor: pointer;
+}
+
+.active_pagination_item {
+  color: #000;
+  text-decoration: underline;
+}
