--- src/styles/posts/post.module.css
@@ -1,4 +1,0 @@
-.post {
-  width: 100%;
-  min-height: calc(100% - 100px);
-}
